+++ src/auth/config.go
@@ -4,7 +4,7 @@ import (
 	apikey "github.com/consensys/quorum-key-manager/src/auth/authenticator/api-key"
 	"github.com/consensys/quorum-key-manager/src/auth/authenticator/oidc"
 	"github.com/consensys/quorum-key-manager/src/auth/authenticator/tls"
-	manifests "github.com/consensys/quorum-key-manager/src/infra/manifests/reader"
+	manifests "github.com/consensys/quorum-key-manager/src/infra/manifests/filesystem"
 )
 
 type Config struct {
