+++ src/infra/manifests/filesystem/config.go
@@ -1,4 +1,4 @@
-package reader
+package filesystem
 
 type Config struct {
 	Path string
