+++ src/nodes/config.go
@@ -1,6 +1,6 @@
 package nodes
 
-import manifests "github.com/consensys/quorum-key-manager/src/infra/manifests/reader"
+import manifests "github.com/consensys/quorum-key-manager/src/infra/manifests/filesystem"
 
 type Config struct {
 	Manifest *manifests.Config
